+++ CHANGELOG.md
@@ -27,6 +27,11 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Improved list view controller and model with filters
 
+#### Bugs
+
+* Fixed potential wrong data-api on ui components
+* Fixed potential wrong trigger on swipe control
+
 ## Version 1.5.1
 ### 2014-03-18